Resources on how to use Flask. I've decided to launch this webapp through Firebase wth Flask.
I've decided to host PokeType through Firebase with Flask. I'm hoping that I could also use Bootstrap, but I'd be satisfied with some normal CSS. I also posted some resources for when I'd need to start developing the back end (flask) and front end. So far it feels like launching a webapp is a MVC design pattern, but I'm still trying to figure out how the pieces work. One thing I haven't found online is the use of sockets and threads, so that's the next thing I should do in terms or webapp hosting research. Prototype is still in development.